% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\name{all_elizabeth_warren}
\alias{all_elizabeth_warren}
\title{All Elizabeth Warren quotes}
\usage{
all_elizabeth_warren()
}
\value{
Character vector, length 356.
}
\description{
All Elizabeth Warren quotes
}
\examples{
all_elizabeth_warren()
}
