% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\name{elizabeth_warran}
\alias{elizabeth_warran}
\title{A random Elizabeth Warren quote}
\usage{
elizabeth_warran(id)
}
\arguments{
\item{id}{a nummeric, will return specific quote if given, otherwise returns
random quote (default).}
}
\value{
Quote printed to console.
}
\description{
A random Elizabeth Warren quote
}
\examples{
elizabeth_warran()

elizabeth_warran(10)
}
